# scramble-openapi-gestionrrhh-v2.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "Gestion RRHH API v2",
        "version": "2.0.0",
        "description": "API de Gestion RRHH (Permisos, Incapacidades, Vacaciones y Novedades)."
    },
    "servers": [
        {
            "url": "https://autogestion.copetran.com.co/gestion/api/v2",
            "description": "Local"
        }
    ],
    "paths": {
...
        "/empleados/incapacidades/radicar": {
            "post": {
                "operationId": "v2.empleados.incapacidades.radicar",
                "tags": [
                    "EmpleadoIncapacidadApi"
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreEmpleadoIncapacidadApiRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "anyOf": [
                                        {
                                            "type": "object",
                                            "properties": {
                                                "ok": {
                                                    "type": "boolean"
                                                },
                                                "message": {
                                                    "type": "string"
                                                },
                                                "data": {
                                                    "type": "object",
                                                    "properties": {
                                                        "id_incapacidad": {
                                                            "type": "string"
                                                        },
                                                        "id_detalle": {
                                                            "type": "string"
                                                        },
                                                        "id_novedad": {
                                                            "type": "string"
                                                        },
                                                        "estado": {
                                                            "type": "string"
                                                        },
                                                        "adjuntos": {
                                                            "type": "object",
                                                            "properties": {
                                                                "guardados": {
                                                                    "type": "integer"
                                                                },
                                                                "items": {
                                                                    "type": "array",
                                                                    "items": {
                                                                        "type": [
                                                                            "object",
                                                                            "null"
                                                                        ],
                                                                        "properties": {
                                                                            "id": {
                                                                                "type": "string"
                                                                            },
                                                                            "id_novedad": {
                                                                                "type": "string"
                                                                            },
                                                                            "id_tipo_documento": {
                                                                                "type": "string"
                                                                            },
                                                                            "tipo_documento": {
                                                                                "type": "string"
                                                                            },
                                                                            "ruta_documento": {
                                                                                "type": "string"
                                                                            },
                                                                            "fuente": {
                                                                                "type": "string"
                                                                            }
                                                                        },
                                                                        "required": [
                                                                            "id",
                                                                            "id_novedad",
                                                                            "id_tipo_documento",
                                                                            "tipo_documento",
                                                                            "ruta_documento",
                                                                            "fuente"
                                                                        ]
                                                                    }
                                                                }
                                                            },
                                                            "required": [
                                                                "guardados",
                                                                "items"
                                                            ]
                                                        }
                                                    },
                                                    "required": [
                                                        "id_incapacidad",
                                                        "id_detalle",
                                                        "id_novedad",
                                                        "estado",
                                                        "adjuntos"
                                                    ]
                                                }
                                            },
                                            "required": [
                                                "ok",
                                                "message",
                                                "data"
                                            ]
                                        },
                                        {
                                            "type": "string"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "422": {
                        "$ref": "#/components/responses/ValidationException"
                    }
                }
            }
        },
...
        "/empleados/permisos/radicar": {
            "post": {
                "operationId": "v2.empleados.permisos.radicar",
                "tags": [
                    "EmpleadoSolicitudApi"
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreEmpleadoPermisoApiRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "anyOf": [
                                        {
                                            "type": "object",
                                            "properties": {
                                                "ok": {
                                                    "type": "boolean"
                                                },
                                                "message": {
                                                    "type": "string"
                                                },
                                                "data": {
                                                    "type": "object",
                                                    "properties": {
                                                        "id_novedad": {
                                                            "type": "string"
                                                        },
                                                        "estado": {
                                                            "type": "string"
                                                        },
                                                        "radicado_por": {
                                                            "type": "object",
                                                            "properties": {
                                                                "documento": {
                                                                    "type": "string"
                                                                },
                                                                "nombre": {
                                                                    "type": "string"
                                                                }
                                                            },
                                                            "required": [
                                                                "documento",
                                                                "nombre"
                                                            ]
                                                        },
                                                        "adjuntos": {
                                                            "type": "object",
                                                            "properties": {
                                                                "ok": {
                                                                    "type": "boolean"
                                                                },
                                                                "guardados": {
                                                                    "type": "integer"
                                                                }
                                                            },
                                                            "required": [
                                                                "ok",
                                                                "guardados"
                                                            ]
                                                        },
                                                        "jefe_directo": {
                                                            "type": [
                                                                "object",
                                                                "null"
                                                            ],
                                                            "properties": {
                                                                "identificacion": {
                                                                    "type": "string"
                                                                },
                                                                "nombre": {
                                                                    "type": "string"
                                                                },
                                                                "correo": {
                                                                    "type": "string"
                                                                }
                                                            },
                                                            "required": [
                                                                "identificacion",
                                                                "nombre",
                                                                "correo"
                                                            ]
                                                        },
                                                        "notificacion_email_enviada": {
                                                            "type": "boolean"
                                                        }
                                                    },
                                                    "required": [
                                                        "id_novedad",
                                                        "estado",
                                                        "radicado_por",
                                                        "adjuntos",
                                                        "jefe_directo",
                                                        "notificacion_email_enviada"
                                                    ]
                                                }
                                            },
                                            "required": [
                                                "ok",
                                                "message",
                                                "data"
                                            ]
                                        },
                                        {
                                            "type": "string"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "422": {
                        "$ref": "#/components/responses/ValidationException"
                    }
                }
            }
        },
...
        "/empleados/vacaciones/radicar": {
            "post": {
                "operationId": "v2.empleados.vacaciones.radicar",
                "tags": [
                    "EmpleadoSolicitudApi"
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreEmpleadoVacacionApiRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "anyOf": [
                                        {
                                            "type": "object",
                                            "properties": {
                                                "ok": {
                                                    "type": "boolean"
                                                },
                                                "message": {
                                                    "type": "string",
                                                    "enum": [
                                                        "Vacaciones radicadas correctamente.",
                                                        "Vacaciones radicadas correctamente. No fue posible enviar la notificacion por correo."
                                                    ]
                                                },
                                                "data": {
                                                    "type": "object",
                                                    "properties": {
                                                        "id_novedad": {
                                                            "type": "string"
                                                        },
                                                        "id_detalle": {
                                                            "type": "string"
                                                        },
                                                        "estado": {
                                                            "type": "string"
                                                        },
                                                        "radicado_por": {
                                                            "type": "object",
                                                            "properties": {
                                                                "documento": {
                                                                    "type": "string"
                                                                },
                                                                "nombre": {
                                                                    "type": "string"
                                                                }
                                                            },
                                                            "required": [
                                                                "documento",
                                                                "nombre"
                                                            ]
                                                        },
                                                        "requiere_aprobacion_rrhh": {
                                                            "type": "boolean"
                                                        },
                                                        "aprobador": {
                                                            "type": [
                                                                "object",
                                                                "null"
                                                            ],
                                                            "properties": {
                                                                "tipo": {
                                                                    "type": "string"
                                                                },
                                                                "documento": {
                                                                    "type": "string"
                                                                },
                                                                "nombre": {
                                                                    "type": "string"
                                                                },
                                                                "correo": {
                                                                    "type": "string"
                                                                }
                                                            },
                                                            "required": [
                                                                "tipo",
                                                                "documento",
                                                                "nombre",
                                                                "correo"
                                                            ]
                                                        },
                                                        "carta": {
                                                            "type": "object",
                                                            "properties": {
                                                                "ok": {
                                                                    "type": "boolean"
                                                                },
                                                                "ruta_documento": {
                                                                    "type": "string"
                                                                },
                                                                "tipo_documento_id": {
                                                                    "type": "string"
                                                                }
                                                            },
                                                            "required": [
                                                                "ok",
                                                                "ruta_documento",
                                                                "tipo_documento_id"
                                                            ]
                                                        },
                                                        "notificacion_email_enviada": {
                                                            "type": "boolean"
                                                        }
                                                    },
                                                    "required": [
                                                        "id_novedad",
                                                        "id_detalle",
                                                        "estado",
                                                        "radicado_por",
                                                        "requiere_aprobacion_rrhh",
                                                        "aprobador",
                                                        "carta",
                                                        "notificacion_email_enviada"
                                                    ]
                                                }
                                            },
                                            "required": [
                                                "ok",
                                                "message",
                                                "data"
                                            ]
                                        },
                                        {
                                            "type": "string"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "422": {
                        "$ref": "#/components/responses/ValidationException"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "StoreEmpleadoIncapacidadApiRequest": {
                "type": "object",
                "properties": {
                    "identificacion": {
                        "type": "string",
                        "maxLength": 50
                    },
                    "causa_id": {
                        "type": "string",
                        "maxLength": 100
                    },
                    "diagnostico_id": {
                        "type": "string",
                        "maxLength": 100
                    },
                    "eps_id": {
                        "type": "string",
                        "maxLength": 100
                    },
                    "arl_id": {
                        "type": "string",
                        "maxLength": 100
                    },
                    "tipo_incapacidad": {
                        "type": "string",
                        "enum": [
                            "INICIO",
                            "PRORROGA"
                        ]
                    },
                    "fecha_inicio": {
                        "type": "string",
                        "format": "date"
                    },
                    "fecha_fin": {
                        "type": "string",
                        "format": "date"
                    },
                    "observacion": {
                        "type": [
                            "string",
                            "null"
                        ],
                        "maxLength": 255
                    },
                    "documento_actor": {
                        "type": "string",
                        "maxLength": 50
                    },
                    "nombre_actor": {
                        "type": [
                            "string",
                            "null"
                        ],
                        "maxLength": 200
                    },
                    "documento_usuario": {
                        "type": [
                            "string",
                            "null"
                        ],
                        "maxLength": 50
                    },
                    "nombre_usuario": {
                        "type": [
                            "string",
                            "null"
                        ],
                        "maxLength": 200
                    },
                    "sistema_origen": {
                        "type": [
                            "string",
                            "null"
                        ],
                        "maxLength": 120
                    },
                    "ip_equipo": {
                        "type": [
                            "string",
                            "null"
                        ],
                        "maxLength": 120
                    },
                    "adjuntos": {
                        "type": [
                            "array",
                            "null"
                        ],
                        "items": {
                            "type": "object",
                            "properties": {
                                "tipo_documento_id": {
                                    "type": [
                                        "string",
                                        "null"
                                    ],
                                    "maxLength": 100
                                },
                                "archivo": {
                                    "type": [
                                        "string",
                                        "null"
                                    ],
                                    "format": "binary",
                                    "contentMediaType": "application/octet-stream",
                                    "maxLength": 10240
                                }
                            }
                        },
                        "maxItems": 8
                    }
                },
                "required": [
                    "identificacion",
                    "causa_id",
                    "diagnostico_id",
                    "eps_id",
                    "arl_id",
                    "tipo_incapacidad",
                    "fecha_inicio",
                    "fecha_fin",
                    "documento_actor"
                ],
                "title": "StoreEmpleadoIncapacidadApiRequest"
            },
...
        },
        "responses": {
            "ValidationException": {
                "description": "Validation error",
                "content": {
                    "application/json": {
                        "schema": {
                            "type": "object",
                            "properties": {
                                "message": {
                                    "type": "string",
                                    "description": "Errors overview."
                                },
                                "errors": {
                                    "type": "object",
                                    "description": "A detailed description of each field that failed validation.",
                                    "additionalProperties": {
                                        "type": "array",
                                        "items": {
                                            "type": "string"
                                        }
                                    }
                                }
                            },
                            "required": [
                                "message",
                                "errors"
                            ]
                        }
                    }
                }
            }
        }
    }
}